Prerender prefetched READMEs
Download the raw README sidecar, render markdown into sanitized HTML,
and let Eleventy embed that HTML into package pages. Packages with a
static README skip the client prefetch and render module.

Share README rendering logic between the browser path and the prerender
script so anchors, alerts, and link rewriting stay consistent.
